+++ src/main/java/com/asu/alliancebank/service/account/IAccountManager.java
@@ -22,7 +22,7 @@ public interface IAccountManager {
 	public abstract List<String> listAllUserLoginIdsFromAccounts(String loggedInUser) throws SQLException;
 	public abstract Map<String, String> getUserListForAddAccount(String loggedInUser); 
 	public abstract List<String> checkDeleteAccountUserIDs(String userLoginIDs[],String loggedInUser)throws SQLException;
-	public abstract Long getAccountBalance(String loggedInUser) throws SQLException;
+	public abstract Double getAccountBalance(String loggedInUser) throws SQLException;
 	public abstract boolean isValidUserLoginID(List<Account> accountList, String userLoginID);
 	public abstract boolean hasAccount(String loggedinId) throws SQLException;
 }
